+++ cpp/oneapi/dal/graph/detail/graph_service_functions_impl.hpp
@@ -18,7 +18,6 @@
 
 #include "oneapi/dal/graph/detail/graph_container.hpp"
 #include "oneapi/dal/graph/graph_common.hpp"
-#include "oneapi/dal/graph/undirected_adjacency_array_graph.hpp"
 
 namespace oneapi::dal::preview::detail {
 
@@ -38,37 +37,4 @@ template <typename Graph>
 ONEAPI_DAL_EXPORT auto get_vertex_neighbors_impl(const Graph &graph,
                                                  const vertex_type<Graph> &vertex) noexcept
     -> const_edge_range_type<Graph>;
-template <typename Graph>
-
-ONEAPI_DAL_EXPORT auto get_vertex_count_impl(const Graph &graph) noexcept
-    -> vertex_size_type<Graph> {
-    const auto &layout = detail::get_impl(graph);
-    return layout->_vertex_count;
-}
-
-template <typename Graph>
-ONEAPI_DAL_EXPORT auto get_edge_count_impl(const Graph &graph) noexcept -> edge_size_type<Graph> {
-    const auto &layout = detail::get_impl(graph);
-    return layout->_edge_count;
-}
-
-template <typename Graph>
-ONEAPI_DAL_EXPORT auto get_vertex_degree_impl(const Graph &graph,
-                                              const vertex_type<Graph> &vertex) noexcept
-    -> edge_size_type<Graph> {
-    const auto &layout = detail::get_impl(graph);
-    return layout->_degrees[vertex];
-}
-
-template <typename Graph>
-ONEAPI_DAL_EXPORT auto get_vertex_neighbors_impl(const Graph &graph,
-                                                 const vertex_type<Graph> &vertex) noexcept
-    -> const_edge_range_type<Graph> {
-    const auto &layout = detail::get_impl(graph);
-    const_edge_iterator_type<Graph> vertex_neighbors_begin =
-        layout->_vertex_neighbors.begin() + layout->_edge_offsets[vertex];
-    const_edge_iterator_type<Graph> vertex_neighbors_end =
-        layout->_vertex_neighbors.begin() + layout->_edge_offsets[vertex + 1];
-    return std::make_pair(vertex_neighbors_begin, vertex_neighbors_end);
-}
 } // namespace oneapi::dal::preview::detail
