+++ cpp/oneapi/dal/graph/undirected_adjacency_array_graph.cpp
@@ -31,4 +31,162 @@ using graph_default = undirected_adjacency_array_graph<empty_value,
                                                        std::int32_t,
                                                        std::allocator<char>>;
 
+template <typename VertexValue,
+          typename EdgeValue,
+          typename GraphValue,
+          typename IndexType,
+          typename Allocator>
+undirected_adjacency_array_graph<VertexValue, EdgeValue, GraphValue, IndexType, Allocator>::
+    undirected_adjacency_array_graph()
+        : impl_(new detail::undirected_adjacency_array_graph_impl<VertexValue,
+                                                                  EdgeValue,
+                                                                  GraphValue,
+                                                                  IndexType,
+                                                                  Allocator>) {}
+
+template <typename VertexValue,
+          typename EdgeValue,
+          typename GraphValue,
+          typename IndexType,
+          typename Allocator>
+undirected_adjacency_array_graph<VertexValue, EdgeValue, GraphValue, IndexType, Allocator>::
+    undirected_adjacency_array_graph(const undirected_adjacency_array_graph &graph)
+        : undirected_adjacency_array_graph() {
+    const auto &layout = detail::get_impl(graph);
+
+    impl_->_vertex_count = layout->_vertex_count;
+    impl_->_edge_count   = layout->_edge_count;
+
+    impl_->_vertex_neighbors = layout->_vertex_neighbors;
+    impl_->_edge_offsets     = layout->_edge_offsets;
+    impl_->_degrees          = layout->_degrees;
+
+    impl_->_vertex_value = layout->_vertex_value;
+    impl_->_edge_value   = layout->_edge_value;
+}
+
+template <typename VertexValue,
+          typename EdgeValue,
+          typename GraphValue,
+          typename IndexType,
+          typename Allocator>
+undirected_adjacency_array_graph<VertexValue, EdgeValue, GraphValue, IndexType, Allocator>::
+    undirected_adjacency_array_graph(undirected_adjacency_array_graph &&graph)
+        : undirected_adjacency_array_graph() {
+    auto &layout = detail::get_impl(graph);
+
+    impl_->_vertex_count  = layout->_vertex_count;
+    layout->_vertex_count = 0;
+
+    impl_->_edge_count  = layout->_edge_count;
+    layout->_edge_count = 0;
+
+    impl_->_vertex_neighbors = std::move(layout->_vertex_neighbors);
+    impl_->_edge_offsets     = std::move(layout->_edge_offsets);
+    impl_->_degrees          = std::move(layout->_degrees);
+
+    impl_->_vertex_value = std::move(layout->_vertex_value);
+    impl_->_edge_value   = std::move(layout->_edge_value);
+}
+
+template <typename VertexValue,
+          typename EdgeValue,
+          typename GraphValue,
+          typename IndexType,
+          typename Allocator>
+undirected_adjacency_array_graph<VertexValue, EdgeValue, GraphValue, IndexType, Allocator>
+    &undirected_adjacency_array_graph<VertexValue, EdgeValue, GraphValue, IndexType, Allocator>::
+    operator=(const undirected_adjacency_array_graph &graph) {
+    if (&graph != this) {
+        const auto &layout = detail::get_impl(graph);
+
+        impl_->_vertex_count = layout->_vertex_count;
+        impl_->_edge_count   = layout->_edge_count;
+
+        impl_->_vertex_neighbors = layout->_vertex_neighbors;
+        impl_->_edge_offsets     = layout->_edge_offsets;
+        impl_->_degrees          = layout->_degrees;
+
+        impl_->_vertex_value = layout->_vertex_value;
+        impl_->_edge_value   = layout->_edge_value;
+    }
+    return *this;
+}
+
+template <typename VertexValue,
+          typename EdgeValue,
+          typename GraphValue,
+          typename IndexType,
+          typename Allocator>
+undirected_adjacency_array_graph<VertexValue, EdgeValue, GraphValue, IndexType, Allocator>
+    &undirected_adjacency_array_graph<VertexValue, EdgeValue, GraphValue, IndexType, Allocator>::
+    operator=(undirected_adjacency_array_graph &&graph) {
+    if (&graph != this) {
+        auto &layout = detail::get_impl(graph);
+
+        impl_->_vertex_count  = layout->_vertex_count;
+        layout->_vertex_count = 0;
+
+        impl_->_edge_count  = layout->_edge_count;
+        layout->_edge_count = 0;
+
+        impl_->_vertex_neighbors = std::move(layout->_vertex_neighbors);
+        impl_->_edge_offsets     = std::move(layout->_edge_offsets);
+        impl_->_degrees          = std::move(layout->_degrees);
+
+        impl_->_vertex_value = std::move(layout->_vertex_value);
+        impl_->_edge_value   = std::move(layout->_edge_value);
+    }
+    return *this;
+}
+
+namespace detail {
+
+template <typename Graph>
+ONEAPI_DAL_EXPORT auto get_vertex_count_impl(const Graph &graph) noexcept
+    -> vertex_size_type<Graph> {
+    const auto &layout = detail::get_impl(graph);
+    return layout->_vertex_count;
+}
+
+template ONEAPI_DAL_EXPORT auto get_vertex_count_impl<graph_default>(
+    const graph_default &graph) noexcept -> vertex_size_type<graph_default>;
+
+template <typename Graph>
+ONEAPI_DAL_EXPORT auto get_edge_count_impl(const Graph &graph) noexcept -> edge_size_type<Graph> {
+    const auto &layout = detail::get_impl(graph);
+    return layout->_edge_count;
+}
+
+template ONEAPI_DAL_EXPORT edge_size_type<graph_default> get_edge_count_impl(
+    const graph_default &graph);
+
+template <typename Graph>
+ONEAPI_DAL_EXPORT auto get_vertex_degree_impl(const Graph &graph,
+                                              const vertex_type<Graph> &vertex) noexcept
+    -> edge_size_type<Graph> {
+    const auto &layout = detail::get_impl(graph);
+    return layout->_degrees[vertex];
+}
+
+template ONEAPI_DAL_EXPORT edge_size_type<graph_default> get_vertex_degree_impl(
+    const graph_default &graph,
+    const vertex_type<graph_default> &vertex);
+
+template <typename Graph>
+ONEAPI_DAL_EXPORT auto get_vertex_neighbors_impl(const Graph &graph,
+                                                 const vertex_type<Graph> &vertex) noexcept
+    -> const_edge_range_type<Graph> {
+    const auto &layout = detail::get_impl(graph);
+    const_edge_iterator_type<Graph> vertex_neighbors_begin =
+        layout->_vertex_neighbors.begin() + layout->_edge_offsets[vertex];
+    const_edge_iterator_type<Graph> vertex_neighbors_end =
+        layout->_vertex_neighbors.begin() + layout->_edge_offsets[vertex + 1];
+    return std::make_pair(vertex_neighbors_begin, vertex_neighbors_end);
+}
+
+template ONEAPI_DAL_EXPORT const_edge_range_type<graph_default> get_vertex_neighbors_impl(
+    const graph_default &graph,
+    const vertex_type<graph_default> &vertex);
+} // namespace detail
 } // namespace oneapi::dal::preview
